% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertnet.R
\name{convertnet}
\alias{convertnet}
\title{convert MRggi results to network data}
\usage{
convertnet(res, fdr.thr = 0.05, c.size = 10)
}
\arguments{
\item{res}{MRggi results}

\item{fdr.thr}{FDR threshold}

\item{c.size}{minimum size of cluster}
}
\value{
network
}
\description{
convert MRggi results to network data
}
