% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSimulData.R
\name{createSimulData}
\alias{createSimulData}
\title{create simulation data set}
\usage{
createSimulData(
  N = 200,
  Bs1g1 = c(0.55, 0.4, 0.25),
  Bs2g2 = c(0.55, 0.4, 0.25),
  Bg1g2 = 0,
  ro = 0
)
}
\arguments{
\item{N}{sample size}

\item{Bs1g1}{effect size between S1 and g1}

\item{Bs2g2}{effect size between S1 and g1}

\item{Bg1g2}{gene-gene interaction between g1 and g2}
}
\value{
res
}
\description{
create simulation data set
}
