% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRggi.R
\name{.multiple_regression}
\alias{.multiple_regression}
\title{performing multiple regression}
\usage{
.multiple_regression(y1, X1)
}
\arguments{
\item{y1}{expression}

\item{X1}{genotype}
}
\value{
beta
}
\description{
performing multiple regression
}
