% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRggi.R
\name{.TSLS}
\alias{.TSLS}
\title{performing two-stage least square}
\usage{
.TSLS(X1, X2, y1, y2)
}
\arguments{
\item{X1}{genotype for gene1}

\item{X2}{genotype for gene2}

\item{y1}{expression for gene1}

\item{y2}{expression for gene2}
}
\value{
stat
}
\description{
performing two-stage least square
}
