% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FineMapping.R
\name{FineMapping}
\alias{FineMapping}
\title{fine-mapping}
\usage{
FineMapping(X, y, L = 10, pip.thr = 0)
}
\arguments{
\item{X}{genotype matrix}

\item{y}{expression matrix}

\item{L}{susie param}

\item{pip.thr}{posterior inclusion probabilities thr}
}
\value{
X.FineMap
}
\description{
fine-mapping
}
