% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRggi.R
\name{MRggi}
\alias{MRggi}
\title{MR-GGI}
\usage{
MRggi(y, X, cor.thr = 0, p.adjust.method = "bonferroni")
}
\arguments{
\item{y}{gene expression; matrix}

\item{X}{cis-SNPs genotype for each gene; list}

\item{cor.thr}{threshold of the correlation between genes}

\item{p.adjust.method}{method for p value adjustment}

\item{Bsg.thr}{threshold of the effect size between cis-Snp set and its gene}
}
\value{
df
}
\description{
MR-GGI
}
